+++ src/app/index.run.js
@@ -1,5 +1,6 @@
-/* global DEVELOPER_MODE ENABLE_LOGGING */
 import { Visualizer } from '@uirouter/visualizer';
+import { states as listingStates } from './pages/listing/listing.state.js';
+import { states as organizationsStates } from './pages/organizations/organizations.state.js';
 
 (function () {
     'use strict';
@@ -9,7 +10,7 @@ import { Visualizer } from '@uirouter/visualizer';
         .run(runBlock);
 
     /** @ngInject */
-    function runBlock ($anchorScroll, $location, $log, $rootScope, $timeout, $transitions, $uiRouter, $window) {
+    function runBlock ($anchorScroll, $location, $log, $rootScope, $timeout, $transitions, $uiRouter, $window, featureFlags) {
 
         // Update page title on state change
         $transitions.onSuccess({}, transition => {
@@ -38,8 +39,24 @@ import { Visualizer } from '@uirouter/visualizer';
             });
         }
 
+        // load states dependent on features
+        if (featureFlags.isOn('listing-edit')) {
+            listingStates['listing-edit-on'].forEach(state => {
+                $uiRouter.stateRegistry.register(state);
+            });
+        } else {
+            listingStates['listing-edit-off'].forEach(state => {
+                $uiRouter.stateRegistry.register(state);
+            });
+        }
+        if (featureFlags.isOn('developer-page')) {
+            organizationsStates.forEach(state => {
+                $uiRouter.stateRegistry.register(state);
+            });
+        }
+
         // Display ui-router state changes
-        if (DEVELOPER_MODE && ENABLE_LOGGING) {
+        if (featureFlags.isOn('states')) {
             $uiRouter.plugin(Visualizer);
         }
     }