+++ src/app/pages/collections/products/products.controller.js
@@ -20,7 +20,7 @@
                 { predicate: 'version', display: 'Version', sortType: 'single' },
                 { predicate: 'decertificationDate', display: 'Date', sortType: 'single', isDate: true },
                 { predicate: 'numMeaningfulUse', display: '# of Known Users', sortType: 'single', nullDisplay: 'Unknown' },
-                { predicate: 'numMeaningfulUseDate', display: 'Last updated date', sortType: 'single', isDate: true, nullDisplay: 'Unknown' },
+                { predicate: 'numMeaningfulUseDate', display: '# Last Updated Date', sortType: 'single', isDate: true, nullDisplay: 'Unknown' },
                 { predicate: 'acb', display: 'ONC-ACB', sortType: 'single' },
                 { predicate: 'chplProductNumber', display: 'CHPL ID', sortType: 'single', sortDefault: true, isLink: true },
                 { predicate: 'certificationStatus', display: 'Status', sortType: 'single' },