+++ src/app/services/network.service.js
@@ -99,7 +99,7 @@ export class NetworkService {
     }
 
     getAcbs (editable) {
-        return this.apiGET('/acbs?editable=' + editable);
+        return this.apiGET('/acbs?editable=' + editable, true);
     }
 
     getAccessibilityStandards () {
@@ -183,7 +183,7 @@ export class NetworkService {
     }
 
     getAtls (editable) {
-        return this.apiGET('/atls?editable=' + editable);
+        return this.apiGET('/atls?editable=' + editable, true);
     }
 
     getCertBodies () {
@@ -220,7 +220,7 @@ export class NetworkService {
         case 'apiDocumentation':
             return this.apiGET('/collections/certified_products?fields=id,edition,developer,product,version,chplProductNumber,certificationStatus,criteriaMet,apiDocumentation,transparencyAttestationUrl');
         case 'bannedDevelopers':
-            return this.apiGET('/decertifications/developers');
+            return this.apiGET('/collections/decertified-developers');
         case 'correctiveAction':
             return this.apiGET('/collections/certified_products?fields=id,edition,developer,product,version,chplProductNumber,certificationStatus,acb,surveillanceCount,openNonconformityCount,closedNonconformityCount');
         case 'decertifiedProducts':