+++ RELEASE_NOTES.md
@@ -1,5 +1,35 @@
 # Release Notes
 
+## Version 10.18.0
+_8 May 2019_
+
+### New Features
+* Use /collections/decertified-developers endpoint instead of deprecated /decertifications/developers.
+* Show "history popup" on all Listings
+* Updated security for edit/split/merge of developers, products, and versions
+* Updated text and column headings on the Decertified Products and Inactive Certificates pages.
+* Updated language on Charts page.
+* Shrink main header navigation on page scroll
+
+### Text changes
+* Tweak text on unsupported browsers page to confirm IE is not supported
+
+---
+
+## Version 10.17.0
+_22 April 2019_
+
+### New features
+* Use new Activity end points
+  * Listing reports
+  * Developer reports
+* Show Developer/Product/Version name changes in Listing history popup
+
+### Bug Fixes
+* Fix behavior of left-side links on the Overview page.
+
+---
+
 ## Version 10.16.0
 _8 April 2019_